+++ app/src/main/java/fr/scootop/data/api/service/ShortlistItemService.kt
@@ -2,17 +2,13 @@ package fr.scootop.data.api.service
 
 import fr.scootop.data.model.ShortlistCollectionItem
 import fr.scootop.data.model.ShortlistItem
-import retrofit2.http.Body
-import retrofit2.http.DELETE
-import retrofit2.http.GET
-import retrofit2.http.POST
-import retrofit2.http.Path
+import retrofit2.http.*
 import rx.Observable
 
 interface ShortlistItemService {
 
     @GET("api/users/{id}/shortlist_items.json")
-    fun getShortlistItems(@Path("id") userId: Int): Observable<List<ShortlistCollectionItem>>
+    fun getShortlistItems(@Path("id") userId: Long): Observable<List<ShortlistCollectionItem>>
 
     @POST("api/shortlist_items.json")
     fun addShortlistItem(@Body shortlistItem: ShortlistItem): Observable<ShortlistItem>
